Add a link to reservation form
